requirements.txt
asgiref==3.2.10
Django==3.0.7
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
django-tables2==2.3.1
django-filter==2.3.0
django-postgres-copy==2.4.1
